[refactor] rename specification dart files with good conventions
