--- apolline-flutter/lib/specifications/operations/LeafSpecification.dart
@@ -1,5 +1,0 @@
-import 'package:apollineflutter/specifications/AbstractCompositeSpecification.dart';
-
-abstract class LeafSpecification extends AbstractCompositeSpecification {
-  bool isSatisfiedBy(Object candidate);
-}
