+++ apolline-flutter/lib/specifications/operations/leaf_specification.dart
@@ -1,0 +1,5 @@
+import 'package:apollineflutter/specifications/abstract_composite_specification.dart';
+
+abstract class LeafSpecification extends AbstractCompositeSpecification {
+  bool isSatisfiedBy(dynamic candidate);
+}
